Refactor: split mentions.ts in 3
